Lista4

+++ src/main/java/org/example/lista1techsieciowe/service/book/BookService.java
@@ -1,7 +1,8 @@
-package org.example.lista1techsieciowe.service;
+package org.example.lista1techsieciowe.service.book;
 
 import org.example.lista1techsieciowe.entity.Book;
 import org.example.lista1techsieciowe.repository.BookRepository;
+import org.example.lista1techsieciowe.service.book.exceptions.BookDoesntExistException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,16 +16,24 @@ public class BookService {
     public BookService(BookRepository bookRepository) {
         this.bookRepository = bookRepository;
     }
+
     public List<Book> getAll() {
         return (List<Book>) bookRepository.findAll();
     }
 
-    public Book getBook(int id) {
-        return bookRepository.findById(id).orElseThrow(() -> new RuntimeException("Book not found"));
+    public Book getBook(Integer id) {
+        return bookRepository.findById(id)
+                .orElseThrow(() -> BookDoesntExistException.create(id));
     }
-    public Book addBook(Book book) { return bookRepository.save(book);
+
+    public Book addBook(Book book) {
+        return bookRepository.save(book);
     }
-    public void deleteBook(int id) {
+
+    public void deleteBook(Integer id) {
+        if (!bookRepository.existsById(id)) {
+            throw BookDoesntExistException.create(id);
+        }
         bookRepository.deleteById(id);
     }
 }
